pyproject.toml
[project]
name = "newsagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=1.14.1",
    "crewai-tools>=1.14.1",
    "ddgs>=9.13.1",
    "duckduckgo-search>=8.1.1",
    "feedparser>=6.0.12",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "litellm>=1.83.0",
    "markdown>=3.10.2",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
]
